+++ src/Common/Result/Transactions/ProcessedTransactionResult.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace B24io\Loyalty\SDK\Services\Admin\Transactions\Result;
+namespace B24io\Loyalty\SDK\Common\Result\Transactions;
 
 use B24io\Loyalty\SDK\Common\Result\Cards\CardItemResult;
 use B24io\Loyalty\SDK\Core\Result\AbstractResult;