+++ src/Services/Admin/Transactions/Transactions.php
@@ -5,12 +5,14 @@ declare(strict_types=1);
 namespace B24io\Loyalty\SDK\Services\Admin\Transactions;
 
 use B24io\Loyalty\SDK\Common\Reason;
+use B24io\Loyalty\SDK\Common\Result\Cards\CardsResult;
+use B24io\Loyalty\SDK\Common\Result\Transactions\ProcessedTransactionResult;
+use B24io\Loyalty\SDK\Common\Result\Transactions\TransactionsResult;
 use B24io\Loyalty\SDK\Common\TransactionType;
 use B24io\Loyalty\SDK\Core\Command;
 use B24io\Loyalty\SDK\Core\Credentials\Context;
+use B24io\Loyalty\SDK\Core\Exceptions\BaseException;
 use B24io\Loyalty\SDK\Services\AbstractService;
-use B24io\Loyalty\SDK\Services\Admin\Transactions\Result\ProcessedTransactionResult;
-use B24io\Loyalty\SDK\Services\Admin\Transactions\Result\TransactionsResult;
 use Fig\Http\Message\RequestMethodInterface;
 use Money\Money;
 use Symfony\Component\Uid\Uuid;
@@ -66,5 +68,23 @@ class Transactions extends AbstractService
                 Uuid::v4()
             )
         ));
+    }
+
+    /**
+     * @throws BaseException
+     */
+    public function count(): int
+    {
+        return (new CardsResult(
+            $this->core->call(
+                new Command(
+                    Context::admin,
+                    RequestMethodInterface::METHOD_GET,
+                    'transactions',
+                    [],
+                    1
+                )
+            )
+        ))->getCoreResponse()->getResponseData()->pagination->total;
     }
 }