+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # loyalty-php-sdk
 
 ## 3.0 (2023.12.15)
-
+* migrate to 'symfony/http-client'
+* bump minimum PHP version requirements to `8.2.*`
+* add `TransactionsFetcher` - fetch transactions
 
 ## 2.1.2 (01.06.2022)
 * bump `guzzlehttp/guzzle` version requirements to `6.*`